+++ packages/showcases/core/csf/components/overlays/modal.csf.js
@@ -1,0 +1,9 @@
+export const parameters = {
+  actions: {
+    handles: ['mouseenter', 'focusin', 'focusout', 'mouseleave'],
+  },
+  design: {
+    type: 'figma',
+    url: 'https://www.figma.com/file/zDZIyayUlr1yTWrsi7cFoo/?node-id=2993%3A12130',
+  },
+};