+++ packages/showcases/core/csf/components/selection-controls/chip.csf.js
@@ -1,0 +1,9 @@
+export const parameters = {
+  actions: {
+    handles: ['mouseenter', 'click', 'focusin', 'focusout'],
+  },
+  design: {
+    type: 'figma',
+    url: 'https://www.figma.com/file/zDZIyayUlr1yTWrsi7cFoo/?node-id=4652%3A11424',
+  },
+};