+++ packages/showcases/css/stories/components/actions/dropdown/dropdown.stories.js
@@ -1,8 +1,11 @@
 import overview from './examples/overview.html';
 
+import { parameters } from '@vtmn/showcase-core/csf/components/actions/dropdown.csf';
+
 export default {
   title: 'Components / Actions / Dropdown',
   argTypes: { onclick: { action: 'clicked' } },
+  parameters,
 };
 
 export const Overview = () => overview;