+++ packages/showcases/css/stories/components/indicators/progressbar/progressbar.stories.js
@@ -5,18 +5,12 @@ import overview from './examples/overview.html';
 import linear from './examples/linear.html';
 import circular from './examples/circular.html';
 
+import { parameters } from '@vtmn/showcase-core/csf/components/indicators/progressbar.csf';
+
 export default {
   title: 'Components / Indicators / Progressbar',
   argTypes: { onchange: { action: 'write' } },
-  parameters: {
-    actions: {
-      handles: ['mouseenter', 'click', 'focusin', 'focusout'],
-    },
-    design: {
-      type: 'figma',
-      url: 'https://www.figma.com/file/zDZIyayUlr1yTWrsi7cFoo/Vitamin---Web-UI-kit-(v0)?node-id=2887%3A11057',
-    },
-  },
+  parameters,
 };
 
 export const Overview = () => overview;