+++ packages/showcases/css/stories/components/navigation/navbar/navbar.stories.js
@@ -1,17 +1,11 @@
 import overview from './examples/overview.html';
 
+import { parameters } from '@vtmn/showcase-core/csf/components/navigation/navbar.csf';
+
 export default {
   title: 'Components / Navigation / Navbar',
   argTypes: { onchange: { action: 'write' } },
-  parameters: {
-    actions: {
-      handles: ['mouseenter', 'click', 'focusin', 'focusout'],
-    },
-    design: {
-      type: 'figma',
-      url: 'https://www.figma.com/file/zDZIyayUlr1yTWrsi7cFoo/Vitamin---Web-UI-kit-(v0)?node-id=4665%3A10809',
-    },
-  },
+  parameters,
 };
 
 export const Overview = () => overview;