+++ packages/showcases/css/stories/components/navigation/tabs/tabs.stories.js
@@ -2,9 +2,12 @@ import overview from './examples/overview.html';
 import small from './examples/small.html';
 import medium from './examples/medium.html';
 
+import { parameters } from '@vtmn/showcase-core/csf/components/navigation/tabs.csf';
+
 export default {
   title: 'Components / Navigation / Tabs',
   argTypes: { onclick: { action: 'clicked' } },
+  parameters,
 };
 
 export const Overview = () => overview;