+++ packages/showcases/css/stories/components/overlays/snackbar/snackbar.stories.js
@@ -3,18 +3,12 @@ import demoScript from './examples/demo.runscript';
 import overview from './examples/overview.html';
 import demo from './examples/demo.html';
 
+import { parameters } from '@vtmn/showcase-core/csf/components/overlays/snackbar.csf';
+
 export default {
   title: 'Components / Overlays / Snackbar',
   argTypes: { onchange: { action: 'write' } },
-  parameters: {
-    actions: {
-      handles: ['mouseenter', 'click', 'focusin', 'focusout'],
-    },
-    design: {
-      type: 'figma',
-      url: 'https://www.figma.com/file/zDZIyayUlr1yTWrsi7cFoo/Vitamin---Web-UI-kit-(v0)?node-id=2796%3A12600',
-    },
-  },
+  parameters,
 };
 
 export const Overview = () => overview;