+++ packages/showcases/css/stories/components/selection-controls/radio-button/radio-button.stories.js
@@ -1,5 +1,6 @@
 import overview from './examples/overview.html';
-import { parameters } from '@vtmn/showcase-core/csf/components/VtmnRadioButton.csf';
+
+import { parameters } from '@vtmn/showcase-core/csf/components/selection-controls/radio-button.csf';
 
 export default {
   title: 'Components / Selection controls / Radio button',