+++ packages/sources/css/tailwind.config.js
@@ -1,16 +1,3 @@
 module.exports = {
   presets: [require('./presets/tailwind')],
-  // Customizations specific to this project would go here
-  theme: {
-    extend: {
-      minHeight: {
-        48: '12rem',
-      },
-    },
-  },
-  variants: {
-    extend: {
-      backgroundColor: ['active'],
-    },
-  },
 };