+++ packages/sources/css/src/components/actions/button/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.4](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-button@0.11.3...@vtmn/css-button@0.11.4) (2022-02-08)
+
+**Note:** Version bump only for package @vtmn/css-button
+
+
+
+
+
+## [0.11.3](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-button@0.11.2...@vtmn/css-button@0.11.3) (2022-02-03)
+
+**Note:** Version bump only for package @vtmn/css-button
+
+
+
+
+
+## [0.11.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-button@0.11.1...@vtmn/css-button@0.11.2) (2022-02-02)
+
+**Note:** Version bump only for package @vtmn/css-button
+
+
+
+
+
+## [0.11.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-button@0.11.0...@vtmn/css-button@0.11.1) (2022-01-28)
+
+**Note:** Version bump only for package @vtmn/css-button
+
+
+
+
+
+# [0.11.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-button@0.10.0...@vtmn/css-button@0.11.0) (2022-01-27)
+
+
+### Features
+
+* **@vtmn/css:** rename hober brand to hover tertiary ([#949](https://github.com/Decathlon/vitamin-web/issues/949)) ([8bbd2ea](https://github.com/Decathlon/vitamin-web/commit/8bbd2ea43e3ccacc10e9a32ab003c629aa125ae1))
+
+
+
+
+
 # [0.10.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-button@0.9.2...@vtmn/css-button@0.10.0) (2022-01-20)