+++ packages/sources/css/src/components/navigation/navbar/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.11](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-navbar@0.1.10...@vtmn/css-navbar@0.1.11) (2022-02-08)
+
+**Note:** Version bump only for package @vtmn/css-navbar
+
+
+
+
+
+## [0.1.10](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-navbar@0.1.9...@vtmn/css-navbar@0.1.10) (2022-02-03)
+
+**Note:** Version bump only for package @vtmn/css-navbar
+
+
+
+
+
+## [0.1.9](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-navbar@0.1.8...@vtmn/css-navbar@0.1.9) (2022-02-02)
+
+**Note:** Version bump only for package @vtmn/css-navbar
+
+
+
+
+
+## [0.1.8](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-navbar@0.1.7...@vtmn/css-navbar@0.1.8) (2022-01-28)
+
+**Note:** Version bump only for package @vtmn/css-navbar
+
+
+
+
+
 ## [0.1.7](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-navbar@0.1.6...@vtmn/css-navbar@0.1.7) (2022-01-20)