+++ packages/sources/css/src/components/selection-controls/toggle/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.4](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-toggle@0.6.3...@vtmn/css-toggle@0.6.4) (2022-02-08)
+
+**Note:** Version bump only for package @vtmn/css-toggle
+
+
+
+
+
+## [0.6.3](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-toggle@0.6.2...@vtmn/css-toggle@0.6.3) (2022-02-03)
+
+**Note:** Version bump only for package @vtmn/css-toggle
+
+
+
+
+
+## [0.6.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-toggle@0.6.1...@vtmn/css-toggle@0.6.2) (2022-02-02)
+
+**Note:** Version bump only for package @vtmn/css-toggle
+
+
+
+
+
+## [0.6.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-toggle@0.6.0...@vtmn/css-toggle@0.6.1) (2022-01-28)
+
+**Note:** Version bump only for package @vtmn/css-toggle
+
+
+
+
+
 # [0.6.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-toggle@0.5.3...@vtmn/css-toggle@0.6.0) (2022-01-20)