+++ packages/sources/css/src/components/structure/skeleton/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.4](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-skeleton@0.3.3...@vtmn/css-skeleton@0.3.4) (2022-02-08)
+
+**Note:** Version bump only for package @vtmn/css-skeleton
+
+
+
+
+
+## [0.3.3](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-skeleton@0.3.2...@vtmn/css-skeleton@0.3.3) (2022-02-03)
+
+**Note:** Version bump only for package @vtmn/css-skeleton
+
+
+
+
+
+## [0.3.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-skeleton@0.3.1...@vtmn/css-skeleton@0.3.2) (2022-02-02)
+
+**Note:** Version bump only for package @vtmn/css-skeleton
+
+
+
+
+
+## [0.3.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-skeleton@0.3.0...@vtmn/css-skeleton@0.3.1) (2022-01-28)
+
+**Note:** Version bump only for package @vtmn/css-skeleton
+
+
+
+
+
 # [0.3.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-skeleton@0.2.9...@vtmn/css-skeleton@0.3.0) (2022-01-20)