+++ packages/sources/css/src/utilities/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.19](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-utilities@0.2.18...@vtmn/css-utilities@0.2.19) (2022-02-08)
+
+**Note:** Version bump only for package @vtmn/css-utilities
+
+
+
+
+
+## [0.2.18](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-utilities@0.2.17...@vtmn/css-utilities@0.2.18) (2022-02-03)
+
+**Note:** Version bump only for package @vtmn/css-utilities
+
+
+
+
+
+## [0.2.17](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-utilities@0.2.16...@vtmn/css-utilities@0.2.17) (2022-02-02)
+
+**Note:** Version bump only for package @vtmn/css-utilities
+
+
+
+
+
+## [0.2.16](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-utilities@0.2.15...@vtmn/css-utilities@0.2.16) (2022-01-28)
+
+**Note:** Version bump only for package @vtmn/css-utilities
+
+
+
+
+
 ## [0.2.15](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-utilities@0.2.14...@vtmn/css-utilities@0.2.15) (2022-01-20)
 
 **Note:** Version bump only for package @vtmn/css-utilities